% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPxplorer-package.R
\docType{package}
\name{CPxplorer-package}
\alias{CPxplorer-package}
\title{CPxplorer: CPxplorer: PCAs analysis and quantification}
\description{
Generates list of ions for mass spectrometric analysis and quantification of PCAs and analogues.
}
\author{
\strong{Maintainer}: Thanh Wang \email{thanh.wang@liu.se} (\href{https://orcid.org/0000-0002-5729-1908}{ORCID})

Authors:
\itemize{
  \item Idoia Beloqui Ezquer \email{idoia.beloqui.ezquer@liu.se}
}

}
\keyword{internal}
