% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPxplorer-package.R
\docType{package}
\name{CPxplorer-package}
\alias{CPxplorer}
\alias{CPxplorer-package}
\title{CPxplorer: CPxplorer: PCAs and analogues analysis and quantification}
\description{
Generates list of ions of PCAs and analogues for mass spectrometric analysis and perform quantification by deconvolution.
}
\author{
\strong{Maintainer}: Thanh Wang \email{thanh.wang@liu.se} (\href{https://orcid.org/0000-0002-5729-1908}{ORCID})

}
\keyword{internal}
